requirements.txt
aerich==0.9.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
click==8.2.1
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pyotp==2.9.0
pypika-tortoise==0.6.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
